requirements.txt
PyAutoGUI==0.9.53
pyperclip==1.8.2
pymongo[srv]==3.12.0
pymongocrypt==1.1.1
selenium
webdriver_manager
docx2txt==0.8
python-docx==0.8.11
steem==1.0.1
ecdsa==0.17.0
steemengine==0.5.3
beem==0.24.26
Markdown==3.3.4
google-api-core==1.31.1
google-api-python-client==2.15.0
google-auth==1.34.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.5
google-cloud-appengine-logging==1.1.0
google-cloud-audit-log==0.2.0
google-cloud-bigquery==2.31.0
google-cloud-core==2.2.1
google-cloud-logging==2.7.0
google-cloud-monitoring==2.8.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
oauth2client
grpcio==1.43.0
elasticsearch==7.15.1
python-logstash-async
PyYAML==6.0
git+https://github.com/jamalex/notion-py.git@refs/pull/352/merge